# Replaced explicit throws with ThrowHelper calls