--- a/src/Hashids.net/ArrayExtensions.cs
+++ b/src/Hashids.net/ArrayExtensions.cs
@@ -1,7 +1,5 @@
 using System;
 using System.Buffers;
-using System.Collections.Generic;
-using System.Text.RegularExpressions;
 
 namespace HashidsNet
 {