Explicitly writing `throw new ...` generates a lot of boilerplate asm code. We can avoid that by using [ThrowHelper](https://docs.microsoft.com/en-us/windows/communitytoolkit/developer-tools/throwhelper) calls.

ASM diffs: https://www.diffchecker.com/yIpy3421

Benchmarks:

Before:
``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19044.1586 (21H2)
AMD Ryzen 5 3600, 1 CPU, 12 logical and 6 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4470.0), X64 RyuJIT
  Job-HBMDUB : .NET 5.0.15 (5.0.1522.11506), X64 RyuJIT
  Job-JTKHOO : .NET 6.0.3 (6.0.322.12309), X64 RyuJIT
  Job-TSPUWP : .NET Framework 4.8 (4.8.4470.0), X64 RyuJIT

OutlierMode=DontRemove  MemoryRandomization=True  

```
|               Method |           Runtime |     Mean |     Error | Code Size | Allocated |
|--------------------- |------------------- |---------:|----------:|----------:|----------:|
|        RoundtripInts |           .NET 5.0 | 4.615 μs | 0.0239 μs |      50 B |     512 B |
|       RoundtripLongs |          .NET 5.0 | 5.113 μs | 0.0441 μs |   2,900 B |     512 B |
|         RoundtripHex |          .NET 5.0 | 4.641 μs | 0.0895 μs |      50 B |   1,344 B |
|         SingleNumber |         .NET 5.0 | 1.166 μs | 0.0092 μs |     153 B |      64 B |
| SingleNumberAsParams |          .NET 5.0 | 1.167 μs | 0.0072 μs |   1,917 B |     160 B |
|                      |            |         |           |           |           |
|        RoundtripInts |          .NET 6.0 | 4.431 μs | 0.0883 μs |      50 B |     512 B |
|       RoundtripLongs |          .NET 6.0 | 4.921 μs | 0.0238 μs |   2,970 B |     512 B |
|         RoundtripHex |          .NET 6.0 | 4.297 μs | 0.0490 μs |      50 B |   1,344 B |
|         SingleNumber |           .NET 6.0 | 1.115 μs | 0.0064 μs |     145 B |      64 B |
| SingleNumberAsParams |           .NET 6.0 | 1.146 μs | 0.0063 μs |   1,902 B |     160 B |
|                      |            |                    |          |           |           |           |
|        RoundtripInts | .NET Framework 4.8 | 5.405 μs | 0.0406 μs |      51 B |     794 B |
|       RoundtripLongs | .NET Framework 4.8 | 6.218 μs | 0.0643 μs |   4,003 B |     907 B |
|         RoundtripHex | .NET Framework 4.8 | 6.964 μs | 0.1363 μs |      52 B |   1,781 B |
|         SingleNumber | .NET Framework 4.8 | 1.507 μs | 0.0134 μs |   2,950 B |     241 B |
| SingleNumberAsParams | .NET Framework 4.8 | 1.474 μs | 0.0175 μs |   2,950 B |     241 B |

After:
``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19044.1586 (21H2)
AMD Ryzen 5 3600, 1 CPU, 12 logical and 6 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4470.0), X64 RyuJIT
  Job-HBMDUB : .NET 5.0.15 (5.0.1522.11506), X64 RyuJIT
  Job-JTKHOO : .NET 6.0.3 (6.0.322.12309), X64 RyuJIT
  Job-TSPUWP : .NET Framework 4.8 (4.8.4470.0), X64 RyuJIT

OutlierMode=DontRemove  MemoryRandomization=True  

```
|               Method |            Runtime |     Mean |     Error | Code Size | Allocated |
|--------------------- |------------------- |---------:|----------:|----------:|----------:|
|        RoundtripInts |           .NET 5.0 | 4.675 μs | 0.0702 μs |      50 B |     512 B |
|       RoundtripLongs |           .NET 5.0 | 5.304 μs | 0.1054 μs |   2,900 B |     512 B |
|         RoundtripHex |           .NET 5.0 | 4.665 μs | 0.0599 μs |      50 B |   1,344 B |
|         SingleNumber |           .NET 5.0 | 1.193 μs | 0.0149 μs |     153 B |      64 B |
| SingleNumberAsParams |          .NET 5.0 | 1.204 μs | 0.0135 μs |   1,917 B |     160 B |
|                      |            |           |           |           |           |
|        RoundtripInts |           .NET 6.0 | 4.452 μs | 0.0357 μs |      50 B |     512 B |
|       RoundtripLongs |           .NET 6.0 | 5.011 μs | 0.0955 μs |   2,970 B |     512 B |
|         RoundtripHex |           .NET 6.0 | 4.325 μs | 0.0499 μs |      50 B |   1,344 B |
|         SingleNumber |          .NET 6.0 | 1.147 μs | 0.0217 μs |     145 B |      64 B |
| SingleNumberAsParams |           .NET 6.0 | 1.144 μs | 0.0086 μs |   1,902 B |     160 B |
|                      |            |          |           |           |           |
|        RoundtripInts | .NET Framework 4.8 | 5.432 μs | 0.0579 μs |      51 B |     794 B |
|       RoundtripLongs | .NET Framework 4.8 | 6.312 μs | 0.0419 μs |   3,621 B |     907 B |
|         RoundtripHex | .NET Framework 4.8 | 7.581 μs | 0.1476 μs |      52 B |   1,781 B |
|         SingleNumber | .NET Framework 4.8 | 1.513 μs | 0.0169 μs |   2,568 B |     241 B |
| SingleNumberAsParams | .NET Framework 4.8 | 1.522 μs | 0.0200 μs |   2,568 B |     241 B |