@pbolduc 2022-03-24
@KeterSCP could you explain further what this PR is doing? I compared the Before and After and in every test, the After is slower in time.  Here is a consolidated view. Perhaps mean time is not what you were after?

![image](https://user-images.githubusercontent.com/1844480/160023259-0900239f-d9aa-4446-94a2-8d0abb3e8d59.png)