@pbolduc 2022-03-25
> > @KeterSCP could you explain further what this PR is doing? I compared the Before and After and in every test, the After is slower in time. Here is a consolidated view. Perhaps mean time is not what you were after? Is it possible you reversed your tables?
> > ![image](https://user-images.githubusercontent.com/1844480/160023259-0900239f-d9aa-4446-94a2-8d0abb3e8d59.png)
> 
> Yes. this PR is not about speed, but about JIT codegen size. If you take a look at `Code Size` column you will see that for .NET Framework the generated asm size is reduced. Also avoiding direct throw, method gets more chances to be inlined. So we should consider such a tradeoff between code size and speed. The most concerning case is `RoundTripHex` on .NET Framework runtime which for my benchmarks in average show regression of about 300 nanoseconds. The rest of the benchmarks are around error bounds.
> 
> For reference you can see [here](https://github.com/dotnet/runtime/blob/3b17fb8c9989790c1ffd76b136ca59502edc60f9/src/libraries/System.Private.CoreLib/src/System/ReadOnlySpan.cs#L65) that the same thing is done in the "official" version of the `ReadOnlySpan<T>`

Ok. Now I understand where you are coming from. Not sure I would agree saving a few KB on code size worth the slightly increased CPU cost (time). Disk is so cheap. I would need a blog post from Stephen Toub to explain why it is beneficial.  Hopefully your reply will provide the project author some things to consider in their decision to accept or not this PR.