@KeterSCP 2022-03-25
@pbolduc 

> Disk is so cheap

Note, that once code was JITed it is stored in the RAM, not on disk