@ullmark 2022-03-25
thank you for the PR! Really nice to see that ppl want to keep improving the library. However I'm not sure if it makes much sense to do this sort of micro optimisations with the cost of a larger code-base? Are you using this library in a way where performance or memory is an issue for you? 

My opinion is that it's our job to optimise JIT code, but we should make the code as maintainable as possible, and make sense for other developers, if there is an issue with the frameworks exception throwing, that should be fixed by those developers.