@KeterSCP 2022-03-25
> with the cost of a larger code-base

Well, here we are talking about **smaller** JIT codegen. Code-base would only grow in the internal ThrowHelper static class, but most of the methods there can be reused. 

> we should make the code as maintainable as possible

As you can see, there are only 4 methods that only throw and IMO it does not complicate maintainability.

> if there is an issue with the frameworks exception throwing

It is not an issue, this is "by design". When you write `throw new ...` there is a need to prepare registers, call exception object ctor, and finally call the internal throw subroutine. All that preparation code lies on cold path in 99% will not be even executed. While having `ThrowHelper`, it is only about making a `call` to that static method if cold path condition is met. JIT will compile this code only once (and only when needed) and all further calls would reuse already compiled code.

> Are you using this library in a way where performance or memory is an issue for you?

@ullmark It's not a big deal for me to be honest, if you think that this PR is not worth it feel free to close, I understand your concerns.