@ullmark 2022-03-25
> It is not an issue, this is "by design". When you write throw new ... there is a need to prepare registers, call exception object ctor, and finally call the internal throw subroutine. All that preparation code lies on cold path in 99% will not be even executed. While having ThrowHelper, it is only about making a call to that static method if cold path condition is met. JIT will compile this code only once (and only when needed) and all further calls would reuse already compiled code.

I understand, I just feel that JIT might not be a valid reason for optimisation since it is generated. In the next .net version, that might not be true any longer. It's like opinions re. javascript loops; You should only use `for` and never `forEach` since it's x ms faster when looping a gazillion items... Does it really matter in your application? I like nice looking code more. :)

But that's just my opinon. @manigandham, since you've been such a kick ass maintainer lately and have been dabbling in the code more lately, it's probably better if you decide 🤘