@manigandham 2022-03-26
Hey @KeterSCP 

Thanks for the PR. My understanding is that the `ThrowHelper` strategy is used to reduce the resulting JIT code. This makes sense for the BCL because that code is used so widely but not so much in a small library like this. 

I re-ran the benchmarks and didn't find a measurable improvement - which leads to the question of what is being saved here? The extra instructions from the `throw` block are very small, especially considering the complexity of the functions, and there's also the downside of altered stack traces when exceptions are actually thrown. 

The null checks in the constructor wouldn't be a bottleneck since its only run once to setup the `Hashids` class, and the `ReadOnlySpan` usage can be replaced with the official .NET `System.Memory` library once we drop `net461` ([going end of life in less than a month](https://devblogs.microsoft.com/dotnet/net-framework-4-5-2-4-6-4-6-1-will-reach-end-of-support-on-april-26-2022/)) so those optimizations can be left to the runtime. 

Given those reasons, I don't think these changes are worth the added complexity. Let me know if you have other thoughts, otherwise I'll close this PR for now. And thanks again for contributing the .NET6 changes with #49!

References:
- https://dunnhq.com/posts/2022/throw-helper/
- https://stackoverflow.com/questions/562479/why-does-sortedlist-implementation-use-throwhelper-instead-of-throwing-directly
- https://stackoverflow.com/questions/1980044/when-should-i-use-a-throwhelper-method-instead-of-throwing-directly
- https://github.com/dotnet/runtime/blob/main/src/libraries/System.Private.CoreLib/src/System/ReadOnlySpan.cs